+++ lib/flop/nimble_schemas.ex
@@ -13,7 +13,7 @@ defmodule Flop.NimbleSchemas do
     default_limit: [type: {:or, [:pos_integer, {:in, [false]}]}, default: 50],
     max_cursor_size: [type: :pos_integer],
     max_filters: [type: {:or, [:pos_integer, {:in, [false]}]}, default: 20],
-    max_limit: [type: :integer, default: 1000],
+    max_limit: [type: {:or, [:pos_integer, {:in, [false]}]}, default: 1000],
     replace_invalid_params: [type: :boolean],
     default_pagination_type: [
       type: {:in, [:offset, :page, :first, :last]},
@@ -67,7 +67,7 @@ defmodule Flop.NimbleSchemas do
       ]
     ],
     default_limit: [type: {:or, [:pos_integer, {:in, [false]}]}],
-    max_limit: [type: :integer],
+    max_limit: [type: {:or, [:pos_integer, {:in, [false]}]}],
     pagination_types: [
       type: {:list, {:in, [:offset, :page, :first, :last]}}
     ],